Update 19.4.2023
*replaced some raw pointer to smart pointer to prevent error
* projectile timer has to be uint64? don't know why but that works LOL :P

+++ Masuguno Tales/src/Projectile.h
@@ -35,8 +35,8 @@ private:
     SDL_Rect srcRect, destRect;
     TransformComponent* mTransform;
     ColliderComponent* mCollider;
-    Uint32 startTime;
-    Uint32 currentTime;
+    Uint64 startTime;
+    Uint64 currentTime;
     std::string tag;
     float damage;
 };
